Now checks for the properties that control the creation and location of undo files.
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk/weka@7528 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ src/main/java/weka/gui/explorer/PreprocessPanel.java
@@ -239,6 +239,11 @@ public class PreprocessPanel
     m_OpenDBBut.setToolTipText("Open a set of instances from a database");
     m_GenerateBut.setToolTipText("Generates artificial data");
     m_UndoBut.setToolTipText("Undo the last change to the dataset");
+    m_UndoBut.setEnabled(ExplorerDefaults.get("enableUndo", "true").equalsIgnoreCase("true"));
+    if (!m_UndoBut.isEnabled()) {
+      m_UndoBut.setToolTipText("Undo is disabled - " +
+      		"see weka.gui.explorer.Explorer.props to enable");
+    }
     m_EditBut.setToolTipText("Open the current dataset in a Viewer for editing");
     m_SaveBut.setToolTipText("Save the working relation to a file");
     m_ApplyFilterBut.setToolTipText("Apply the current filter to the data");
@@ -1231,11 +1236,34 @@ public class PreprocessPanel
    * @throws Exception 	if an error occurs
    */
   public void addUndoPoint() throws Exception {
+    if (!ExplorerDefaults.get("enableUndo", "true").equalsIgnoreCase("true")) {
+      return;
+    }
     
     if (m_Instances != null) {
       // create temporary file
       File tempFile = File.createTempFile("weka", SerializedInstancesLoader.FILE_EXTENSION);
       tempFile.deleteOnExit();
+      if (!ExplorerDefaults.get("undoDirectory", "%t").equalsIgnoreCase("%t")) {
+        String dir = ExplorerDefaults.get("undoDirectory", "%t");
+        File undoDir = new File(dir);
+        if (undoDir.exists()) {
+          String fileName = tempFile.getName();
+          File newFile = new File(dir + File.separator + fileName);
+          if (undoDir.canWrite()) {
+            newFile.deleteOnExit();
+            tempFile = newFile;
+          } else {
+            System.err.println("Explorer: it doesn't look like we have permission" +
+            		" to write to the user-specified undo directory " +
+            		"'" + dir + "'");
+          }
+        } else {
+          System.err.println("Explorer: user-specified undo directory '" +
+              dir + "' does not exist!");
+        }
+      }
+    
 
       ObjectOutputStream oos = 
 	new ObjectOutputStream(
